mcp-server/requirements.txt
# MCP Server Dependencies
mcp>=1.0.0
atlassian-python-api>=3.41.0
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.27.0
